Refactor reusable pandas functions into separate file
